redux removal in progress part 2
