--- app/src/marketplace/modules/transactions/constants.ts
@@ -1,2 +1,0 @@
-export const ADD_TRANSACTION = 'marketplace/transactions/ADD_TRANSACTION'
-export const COMPLETE_TRANSACTION = 'marketplace/transactions/COMPLETE_TRANSACTION'
