--- app/src/shared/components/Notifications/BasicNotification/basic.pcss
@@ -1,32 +1,0 @@
-.container {
-  line-height: 28px;
-  display: flex;
-  align-items: flex-start;
-}
-
-.title {
-  font-size: 16px;
-  line-height: 20px;
-  font-weight: 500;
-  color: #323232;
-}
-
-.description {
-  font-size: 14px;
-  color: #323232;
-  margin-top: 5px;
-}
-
-.icon {
-  margin-right: 16px;
-}
-
-.iconSize {
-  width: 20px;
-  height: 20px;
-}
-
-.textBlock {
-  display: flex;
-  flex-direction: column;
-}
