--- app/src/shared/components/Notifications/CustomNotification/index.tsx
@@ -1,16 +1,0 @@
-import React, { FunctionComponent, ReactNode } from 'react'
-import styles from '../BasicNotification/basic.pcss'
-
-type Props = {
-    children?: ReactNode
-}
-
-const CustomNotification: FunctionComponent<Props> = ({ children }: Props) => (
-    <div className={styles.container}>
-        <div className={styles.textBlock}>
-            {children}
-        </div>
-    </div>
-)
-
-export default CustomNotification
