--- app/src/shared/hooks/useBalances.ts
@@ -1,26 +1,0 @@
-import { useCallback, useMemo } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { selectBalances } from '$shared/modules/user/selectors'
-import { updateBalances } from '$shared/modules/user/actions'
-import { useWalletAccount } from '$shared/stores/wallet'
-
-export function useBalances() {
-    const account = useWalletAccount()
-
-    const dispatch = useDispatch()
-    // Balances are stored in redux state and polled regularly in GlobalInfoWatcher component.
-    const balances = useSelector(selectBalances)
-    // Fetches & updates account balances to redux state
-    const update = useCallback(() => {
-        if (account) {
-            dispatch(updateBalances(account) as any) // quick & dirty typing patch
-        }
-    }, [dispatch, account])
-    return useMemo(
-        () => ({
-            update,
-            balances,
-        }),
-        [update, balances],
-    )
-}
