--- app/src/shared/modules/user/constants.ts
@@ -1,1 +1,0 @@
-export const SET_BALANCE = 'shared/user/SET_BALANCE'
