+++ app/src/shared/stores/wallet.ts
@@ -4,8 +4,8 @@ import detectProvider from '@metamask/detect-provider'
 import { create } from 'zustand'
 import Web3 from 'web3'
 import { MetaMaskInpageProvider } from '@metamask/providers'
-import { lookupEnsName } from '$shared/modules/user/services'
 import { isEthereumAddress } from '$app/src/marketplace/utils/validate'
+import {lookupEnsName} from "$shared/utils/lookupEnsName"
 
 interface RequestArguments {
     readonly method: string
