+++ app/src/shared/utils/constants.ts
